requirements.txt
selenium==4.1.5
python-dotenv==0.20.0